+++ BeyondStorage/Source/Configuration/ConfigLocalisation.cs
@@ -4,7 +4,7 @@ internal static class ConfigLocalisation
 {
     internal const string RANGE_SETTING_TOOLTIP_KEY = "bsSettingRangeTooltip";
     internal const string ALLOW_PUSH_TO_ALLIED_VEHICLES_SETTING_TOOLTIP_KEY = "bsSettingAllowPushToAlliedVehiclesToolTip";
-    internal const string CONSUME_FROM_DRONES_SETTING_TOOLTIP_KEY = "bsSettingConsumeFromDronesTooltip";
-    internal const string CONSUME_FROM_VEHICLES_TOOLTIP_KEY = "bsSettingConsumeFromVehiclesTooltip";
+    internal const string INCLUDE_DRONES_SETTING_TOOLTIP_KEY = "bsSettingIncludeDronesInNetworkToolTip";
+    internal const string INCLUDE_VEHICLES_TOOLTIP_KEY = "bsSettingIncludeVehiclesInNetworkToolTip";
     internal const string IS_DEBUG_LOGGING_SETTING_TOOLTIP_KEY = "bsSettingIsDebugTooltip";
 }